Refactoring
1. Changed the test matrix of coefficients of a system of linear equations. The new matrix is ​​singular and has a dimension of 10 × 10.

2. Added the writing of the absolute error of the solution found.

3. The writing of the decision and the absolute error is corrected in a scientific format.

4. The accuracy of the calculations is corrected for double.

+++ LUDecomposition/LUDecomposition/Matrix.h
@@ -3,8 +3,8 @@
 #include <fstream>
 #include <iostream>
 
-typedef float real;
-typedef float realScal;
+typedef double real;
+typedef double realScal;
 
 void InputSize(int&, int&);													// Entering A, AL, AU dimensions
 void Input(int&, int&, real *, real *, real *, real *, int *);					// Entering matrix A in profile format
